update stories [skip ci]
